feat(ARCO-212): Rename folders to camel case
