+++ pkg/api/handler/internal/merkle_verifier/merkle_verifier_adapter.go
@@ -1,4 +1,4 @@
-package merkleverifier
+package merkle_verifier
 
 import (
 	"context"
